cancel added to modal uploadmedia v2

+++ database/migrations/2025_12_04_152923_create_media_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('mime_type');
             $table->unsignedBigInteger('size');
-            $table->float('duration_seconds')->nullable();
+            $table->unsignedBigInteger('duration_seconds')->nullable();
             $table->string('checksum')->nullable();
             $table->foreignId('created_by')->constrained('users');
             $table->timestamps();
